@@ -1,32 +1,25 @@
 package com.androidtutz.anushka.tmdbclient.view;
 
-import android.arch.lifecycle.Observer;
-import android.arch.lifecycle.ViewModelProviders;
+import androidx.lifecycle.Observer;
+import androidx.lifecycle.ViewModelProviders;
 import android.content.res.Configuration;
-import android.databinding.DataBindingUtil;
-import android.support.annotation.Nullable;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AppCompatActivity;
+import androidx.databinding.DataBindingUtil;
+import androidx.annotation.Nullable;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.DefaultItemAnimator;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.DefaultItemAnimator;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.androidtutz.anushka.tmdbclient.R;
 import com.androidtutz.anushka.tmdbclient.adapter.MovieAdapter;
 import com.androidtutz.anushka.tmdbclient.databinding.ActivityMainBinding;
 import com.androidtutz.anushka.tmdbclient.model.Movie;
-import com.androidtutz.anushka.tmdbclient.model.MovieDBResponse;
-import com.androidtutz.anushka.tmdbclient.service.MovieDataService;
-import com.androidtutz.anushka.tmdbclient.service.RetrofitInstance;
 import com.androidtutz.anushka.tmdbclient.viewmodel.MainActivityViewModel;
 
 import java.util.ArrayList;
 import java.util.List;
-
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
 
 public class MainActivity extends AppCompatActivity {
 
